Issue #30: Add a script to handle initial data contribution
This script is the basis for contributing data. It is intended for an
initial kick-off contribution. Other options might be added, but
real-time contribution (status updates, etc) would be done by a separate
tool/daemon that will be notified by Koha of the changes it needs to
propagate to the central servers.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>