Issue #97: Have itemhold create a patron for borrowing site if required
This patch introduces a new configuration entry: 'partners_library_id'
which will be used as the home library for the created patrons.

This patrons will represent the borrowing site and their existence will
be checked when itemhold is requested, and created if required. A
follow-up issue should take care of using the introduced API for
retrieving the agencies descriptions..

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>