Issue #97: Add sync_agencies.pl script
This script is designed to be run on initializing the configuration. It
will take care of creating Koha patrons for each consortia member
library and track it on a specilized plugin table.

This script is also designed to be run as a cronjob. The current
itemhold implementation doesn't create patrons on-the-fly so having this
run as a cronjob is advised.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>