Issue #101: PUT /innreach/v2/circ/transferrequest support
This patch introduces initial support for transferrequest on the
borrowing site.

It is a simple endpoint that updates item and request information in the
event of a transferrequest API request.

TODO: Maybe store the 'old' itemId and something highlighting the
transfer took place. Koha::Illrequestattributes are not repeatable as is
(PRIMARY KEY (illrequest_id,type).

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>