Issue #109: Refactor the code to handle multiple central servers
The original code was close to the goal, but OAuth2 configuration needs
to be per-server, and also the task queue needs fixes as some central
servers might fail to process the request and others not (specially for
record contribution).

This patch still lacks proper per-central server handling on the task
queue.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>