Issue #153: Add central item type to local mapping
In order to be able to define different loan periods for each central
item type, we need to be able to define central server material types
mappings to local ones.

This patch adds that and the README is updated to reflect that.

The original _default_item_type_ configuration entry will still be used
when appropriate (i.e. as a fallback value).

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>