Issue #177: Implement the 'receiveunshipped' notification step [B]
For the borrowing site, the spec has a 'receiveupshipped' message that
can be sent to the INN-Reach central server in order to make it aware we
never received teh 'itemshipped' message.

This patch implements such step, and a form to enter teh call number and
barcode for the item. It will finally implement the same functionality
as the 'itemshipped' endpoint controller, but user generated.

I took the chance to refactor things a bit for reusability.

More refactorings to come.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>