Issue #261: Remove Class::Accessor and simplify Contribution.pm
Contribution.pm incorrectly initializes all the configured OAuth2 user agents for all the central servers, even if being used for something else. This leads to situations in which some internal checks might fail because of temporary INNReach connectivity issues.
This patch aligns the contribution code to the rest of the plugin codebase, using $plugin->get_ua() when really needed instead.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>