Issue #289: Prepare DB for timed tasks
This patch adds a column for adding a timestamp to be used for datetime
comparisson. This way, the task queue job can require jobs to be
scheduled to be run with `run_after >= NOW()`

The patch also adds the current plugin background jobs as job types to
prepare the ground for migrating them.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>