[#297] Implement the 'notices_content' hook
This patch implements the hook. It only does something when called in a
`HOLD_SLIP` context. Other cases can be added later.

It makes available the following keys for using in the `HOLD_SLIP`
letter:

* [% plugin_content.innreach.ill_request | html %]
* [% plugin_content.innreach.itemId | html %]
* [% plugin_content.innreach.pickupLocation | html %]
* [% plugin_content.innreach.patronName | html %]
* [% plugin_content.innreach.centralPatronType | html %]

The key `ill_request` is the full `Koha::ILL::Request` object. If the
hold is not linked to an ILL request, then nothing is injected. So it's
a good key to check in the template logic to detect the ILL use case.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>